requirements.txt
e2b-code-interpreter==2.2.0
gradio>=5.49.0
notebook>=7.4.4
openai==2.4.0
pillow>=11.3.0
python-dotenv>=1.1.1
rich>=14.1.0
tiktoken>=0.9.0
gradio-browser==0.0.3
gradio-aicontext==0.0.4
fuzzywuzzy==0.18.0
ipython>=8.0.0